+++ src/continuity/__init__.py
@@ -5,13 +5,21 @@ The package is structured into the following modules:
 
 ::cards:: cols=2
 
+- title: Benchmarks
+  content: Benchmarks for testing operator architectures.
+  url: benchmarks/index.md
+
+- title: Data
+  content: Data sets and data utility functions.
+  url: data/index.md
+
+- title: Discrete
+  content: Discretization utilities like samplers.
+  url: discrete/index.md
+
 - title: Operators
   content: Neural operator implementations.
   url: operators/index.md
-
-- title: Data
-  content: Data sets for training.
-  url: data/index.md
 
 - title: PDE
   content: Loss functions for physics-informed training.
@@ -20,6 +28,14 @@ The package is structured into the following modules:
 - title: Plotting
   content: Plotting utilities.
   url: plotting/index.md
+
+- title: Trainer
+  content: Default training loop for operator models.
+  url: trainer/index.md
+
+- title: Transforms
+  content: Transformations for operator inputs and outputs.
+  url: transform/index.md
 
 ::/cards::